@@ -1,8 +1,9 @@
-from datetime import datetime
+from datetime import datetime, timezone
 
-from dynamodb_wrapper_V2.dynamodb_wrapper.models import PipelineConfig, PipelineRunLog, TableConfig
-from dynamodb_wrapper_V2.dynamodb_wrapper.models.pipeline_run_log import LogLevel, RunStatus
-from dynamodb_wrapper_V2.dynamodb_wrapper.models.table_config import DataFormat, TableType
+from dynamodb_wrapper.models.domain_models import (
+    PipelineConfig, PipelineRunLog, TableConfig,
+    LogLevel, RunStatus, DataFormat, TableType
+)
 
 
 class TestPipelineConfig:
@@ -160,12 +161,12 @@ class TestPipelineRunLog:
 
     def test_pipeline_run_log_with_stages(self):
         """Test pipeline run log with stage information."""
-        from dynamodb_wrapper_V2.dynamodb_wrapper.models.pipeline_run_log import StageInfo
+        from dynamodb_wrapper.models.domain_models import StageInfo
 
         stage = StageInfo(
             stage_name="extract",
             status=RunStatus.SUCCESS,
-            start_time=datetime.utcnow(),
+            start_time=datetime.now(timezone.utc),
             records_processed=1000
         )
 
@@ -186,7 +187,7 @@ class TestPipelineRunLog:
 
     def test_pipeline_run_log_with_data_quality(self):
         """Test pipeline run log with data quality results."""
-        from dynamodb_wrapper_V2.dynamodb_wrapper.models.pipeline_run_log import DataQualityResult
+        from dynamodb_wrapper.models.domain_models import DataQualityResult
 
         dq_result = DataQualityResult(
             check_name="row_count",
