@@ -2,10 +2,10 @@
 """
 DynamoDB Wrapper Library V2
 
-The next generation DynamoDB wrapper library with enhanced features and improvements.
-A comprehensive Python library that provides a repository pattern wrapper 
-around AWS DynamoDB using boto3 and Pydantic, specifically designed for 
-PySpark pipeline environments.
+The next generation DynamoDB wrapper library with CQRS architecture and model-driven key management.
+A comprehensive Python library that provides a CQRS (Command Query Responsibility Segregation) 
+wrapper around AWS DynamoDB using boto3 and Pydantic, with type-safe domain operations 
+and automated key management.
 """
 
 __version__ = "1.0.0"
