@@ -1,11 +1,25 @@
+# Base exception class
 from .base import DynamoDBWrapperError
-from .connection import ConnectionError
-from .item_not_found import ItemNotFoundError
-from .validation import ValidationError
+
+# Domain-specific exceptions (consolidated from 6 separate files)
+from .domain_exceptions import (
+    ValidationError,
+    ItemNotFoundError,
+    NotFoundError,
+    ConflictError,
+    ConnectionError,
+    RetryableError,
+)
 
 __all__ = [
+    # Base exception
     "DynamoDBWrapperError",
+    
+    # Domain exceptions (alphabetically ordered)
+    "ConflictError",
+    "ConnectionError", 
     "ItemNotFoundError",
+    "NotFoundError",
+    "RetryableError",
     "ValidationError",
-    "ConnectionError",
 ]
